abstract into web-monetization.js
